refactor(torghut): remove autonomy gate dynamic wrappers
